+++ packages/internal/utils/src/export/partition.js
@@ -1,0 +1,9 @@
+export default (list, condition) => {
+  return list.reduce(
+    (result, element) => {
+      result[condition(element) ? 0 : 1].push(element)
+      return result
+    },
+    [[], []]
+  )
+}