+++ packages/internal/base/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="8.22.0"></a>
+# [8.22.0](https://bitbucket.org/stacc-flow/bento/compare/v8.21.0...v8.22.0) (2018-11-16)
+
+
+### Features
+
+* **base:** make slider handle styleable ([cfb79b1](https://bitbucket.org/stacc-flow/bento/commits/cfb79b1))
+* **fundu-theme:** styling of api-management ([86c033c](https://bitbucket.org/stacc-flow/bento/commits/86c033c))
+* **fundu-theme:** styling of devportal menu, header, search, article ++ ([9a89b75](https://bitbucket.org/stacc-flow/bento/commits/9a89b75))
+
+
+
+
+
 <a name="8.19.0"></a>
 # [8.19.0](https://bitbucket.org/stacc-flow/bento/compare/v8.18.0...v8.19.0) (2018-11-14)