+++ packages/internal/mr-x/CHANGELOG.md
@@ -3,6 +3,27 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="8.22.0"></a>
+# [8.22.0](https://bitbucket.org/stacc-flow/bento/compare/v8.21.0...v8.22.0) (2018-11-16)
+
+**Note:** Version bump only for package mr-x
+
+
+
+
+
+<a name="8.20.0"></a>
+# [8.20.0](https://bitbucket.org/stacc-flow/bento/compare/v8.19.0...v8.20.0) (2018-11-15)
+
+
+### Features
+
+* **mr-x:** parse text and link_names ([5f11484](https://bitbucket.org/stacc-flow/bento/commits/5f11484))
+
+
+
+
+
 <a name="8.19.0"></a>
 # [8.19.0](https://bitbucket.org/stacc-flow/bento/compare/v8.18.0...v8.19.0) (2018-11-14)