+++ packages/internal/release-bento/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="8.22.0"></a>
+# [8.22.0](https://bitbucket.org/stacc-flow/bento/compare/v8.21.0...v8.22.0) (2018-11-16)
+
+**Note:** Version bump only for package @staccx/release-bento
+
+
+
+
+
+<a name="8.21.0"></a>
+# [8.21.0](https://bitbucket.org/stacc-flow/bento/compare/v8.20.0...v8.21.0) (2018-11-15)
+
+
+### Features
+
+* **release-bento:** add scope to ls after release ([bc25ba8](https://bitbucket.org/stacc-flow/bento/commits/bc25ba8))
+
+
+
+
+
+<a name="8.20.0"></a>
+# [8.20.0](https://bitbucket.org/stacc-flow/bento/compare/v8.19.0...v8.20.0) (2018-11-15)
+
+
+### Bug Fixes
+
+* **release-bento:** add link to changelog in success output ([440d3c8](https://bitbucket.org/stacc-flow/bento/commits/440d3c8))
+
+
+
+
+
 <a name="8.19.0"></a>
 # [8.19.0](https://bitbucket.org/stacc-flow/bento/compare/v8.18.0...v8.19.0) (2018-11-14)